% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadRedcapReport.R
\name{downloadRedcapReport}
\alias{downloadRedcapReport}
\title{Download RedCap Report}
\arguments{
\item{redcapTokenName}{A string representing the name of the API token.}

\item{redcapUrl}{A string representing the URL for RedCap.}

\item{redcapReportId}{A string representing the report ID.}
}
\value{
A tibble containing the report data.
}
\description{
Securely downloads a RedCap report using an API token from the user's .REnviron file.
}
