% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikBernoulli.R
\name{logLikBernoulli}
\alias{logLikBernoulli}
\title{Log-Likelihood Bernoulli}
\usage{
logLikBernoulli(data)
}
\arguments{
\item{data}{A vector of 0s and 1s.}
}
\value{
A numeric value representing the probability \( p \) that maximizes the log-likelihood.
}
\description{
Calculates the probability \( p \) that maximizes the log-likelihood for a given vector of binary data.
}
