% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimumN.R
\name{minimumN}
\alias{minimumN}
\title{Minimum N}
\arguments{
\item{x1}{A numeric vector for one sample.}

\item{x2}{An optional numeric vector for a second sample.}
}
\value{
A pwr object containing the necessary sample size.
}
\description{
Uses pwr::pwr.t2n.test to determine the minimum sample size needed for a t-test.
}
