% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcApprox.R
\name{pcApprox}
\alias{pcApprox}
\title{Principal Component Approximation}
\usage{
pcApprox(x, npc)
}
\arguments{
\item{x}{A numeric matrix or dataframe.}

\item{npc}{The number of principal components to use.}
}
\value{
An approximated dataset.
}
\description{
Returns an approximation to a dataset based on a given number of principal components.
}
