% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeNames.R
\name{standardizeNames}
\alias{standardizeNames}
\title{Standardize Names}
\arguments{
\item{data}{A tibble.}
}
\value{
A tibble with standardized names.
}
\description{
Standardizes the variable names of a dataset.
}
