% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survCurv.R
\name{survCurv}
\alias{survCurv}
\title{Survival Curve}
\arguments{
\item{status}{A numerical vector indicating the survival status.}

\item{time}{A numerical vector indicating the time to event.}
}
\value{
A survival curve plot.
}
\description{
Calculates and plots a survival curve \( S(t) \) using the survival package.
}
