% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscale.R
\name{unscale}
\alias{unscale}
\title{Unscale}
\usage{
unscale(x)
}
\arguments{
\item{x}{A scaled numeric vector.}
}
\value{
A numeric vector with the original values.
}
\description{
Reverses the scaling effect applied to a numeric vector.
}
