% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{downloadRedcapReport}
\alias{downloadRedcapReport}
\title{Download a Report from REDCap}
\usage{
downloadRedcapReport(redcapTokenName, redcapUrl, redcapReportId)
}
\arguments{
\item{redcapTokenName}{A string specifying the name of the environment variable where
the REDCap API token is stored. This token is used to authenticate the request.}

\item{redcapUrl}{A string containing the base URL of the REDCap instance.}

\item{redcapReportId}{An integer or string specifying the unique identifier of the report to be downloaded.}
}
\value{
Returns a tibble of the requested report.
}
\description{
This function downloads a specified report from a REDCap project using the REDCap API.
The report is requested via POST request, retrieved in CSV format, and returned as a tibble.
}
\examples{
# Assuming you have a REDCap token stored in an environment variable named "REDCAP_TOKEN"
# and the REDCap instance URL is "https://redcap.example.com/api/", for a report with ID 12345:
downloadRedcapReport("REDCAP_TOKEN", "https://redcap.example.com/api/", "12345")

}
