% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{logLikBernoulli}
\alias{logLikBernoulli}
\title{Calculate the Maximum Likelihood Estimate of Probability for a Bernoulli Distribution}
\usage{
logLikBernoulli(data)
}
\arguments{
\item{data}{A binary vector (0s and 1s) representing the trial outcomes.}
}
\value{
The value of \code{p} that maximizes the Bernoulli log-likelihood for the provided data.
}
\description{
This function calculates the maximum likelihood estimate (MLE) of the probability parameter \code{p}
for a Bernoulli distribution given a binary data vector. It searches across a grid of potential
\code{p} values ranging from 0.001 to 0.999 in steps of 0.001.
}
\examples{
sample_data = c(1, 0, 0, 0, 1, 1, 1)
mle_p = logLikBernoulli(sample_data)

}
