% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{minimumN}
\alias{minimumN}
\title{Calculate Minimum Sample Size for Power Analysis}
\usage{
minimumN(x1, x2 = NULL, alpha = 0.05, power = 0.8)
}
\arguments{
\item{x1}{Numeric vector for the first group or the single sample.}

\item{x2}{Optional numeric vector of data for the second sample (default is \code{NULL} for one-sample tests).}

\item{alpha}{Desired significance level (default is 0.05).}

\item{power}{Desired power of the test (default is 0.80).}
}
\value{
The ceiling of the calculated minimum sample size necessary to achieve the specified power
for the given significance level and effect size.
}
\description{
This function calculates the minimum sample size required to achieve a given power for a one-sample
or two-sample t-test. For a one-sample test, the null hypothesis is that the mean of \code{x1} is zero.
For a two-sample test, the null hypothesis is that the means of \code{x1} and \code{x2} are equal.
}
\examples{
# For a one-sample test
x1 <- rnorm(100, 1, 1)
minimumN(x1)

# For a two-sample test
x2 <- rnorm(100, 1.2, 1)
minimumN(x1, x2)

}
