% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pcApprox}
\alias{pcApprox}
\title{Perform PCA and Approximate Original Data}
\usage{
pcApprox(x, npc)
}
\arguments{
\item{x}{A numeric matrix of data to be analyzed.}

\item{npc}{The number of principal components to use for the approximation.}
}
\value{
A matrix that represents an approximation of the original data,
reconstructed using the specified number of principal components and
rescaled to match the scale and center of the original data.
}
\description{
This function scales and centers the data, performs PCA, and then reconstructs
an approximation of the original data using a specified number of principal
components. The data is then rescaled and recentered to match the original scale.
}
\examples{
test_data = matrix(rnorm(100), nrow = 10, ncol = 10)
approx_data = pcApprox(test_data, npc = 2)
print(approx_data)

}
