% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{standardizeNames}
\alias{standardizeNames}
\title{Standardize Column Names to a Specific Case}
\usage{
standardizeNames(data, target_case = "small_camel")
}
\arguments{
\item{data}{A dataframe or tibble whose column names are to be standardized.}

\item{target_case}{A string specifying the target case format for the column names.
The default is "small_camel". Supports all case types supported by
\code{snakecase::to_any_case} including "snake", "small_camel", "big_camel",
and "screaming_snake".}
}
\value{
A dataframe or tibble with standardized column names.
}
\description{
This function standardizes the column names of a dataframe to a specific naming convention.
It uses \code{janitor::make_clean_names} and \code{snakecase::to_any_case} to clean columns names and convert them
to the specified case format. The case format can be adjusted via the \code{target_case} parameter.
}
\examples{
data = data.frame(`First name` = c("Alice", "Bob"), `Last.Name` = c("Smith", "Jones"))
new_data = standardizeNames(data, target_case = "snake")

}
