% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{survCurv}
\alias{survCurv}
\title{Estimate and Plot a Survival Curve}
\usage{
survCurv(status, time)
}
\arguments{
\item{status}{A numeric vector where '1' indicates an event (e.g., death) and '0' indicates censoring.}

\item{time}{A numeric vector representing the time at which each event or censoring occurred.}
}
\value{
A ggplot object displaying the estimated survival curve over time.
}
\description{
This function creates a survival curve based on status and time data provided. It first formats the data into a
tibble, calculates cumulative events and censorships, and then computes the survival probability at each unique
time unit. Finally, it plots the resulting survival curve using ggplot2.
}
\examples{
status_vector <- c(1, 0, 1, 0, 0, 1)
time_vector <- c(5, 12, 15, 20, 22, 25)
survival_plot = survCurv(status_vector, time_vector)
print(survival_plot)

}
