% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{unscale}
\alias{unscale}
\title{Undo Scaling and Centering of Data}
\usage{
unscale(x)
}
\arguments{
\item{x}{A numeric vector or matrix that has been scaled and/or centered.}
}
\value{
The original data with scaling and centering transformations reversed.
If no scaling or centering has been done, the input data \code{x} is returned unchanged.
}
\description{
This function reverts any scaling and centering transformations applied to data.
It checks for attributes indicating scaling and centering transformations ('scaled:center'
and 'scaled:scale') and applies the inverse operations to restore the original data values.
}
\examples{
# Scale and then unscale data
data = 1:10
scaled_data = scale(data)
original_data = unscale(scaled_data)

}
